OneDrive Known Folder Move Shortcut De-Duplicator
Script to monitor and remove duplicate desktop shortcuts created when OneDrive Known Folder Move initialises

This is work in progress. Please offer suggestons, all welcome. My hope is to run this as a Scheduled Task at user logon to react when OneDrive attemtps to create "copies" of application shorcuts on the desktop.

Next Commit will allow parameters for timers and filters